migrate TestHandleNewValidator to new simapp
