+++ x/slashing/keeper/common_test.go
@@ -1,0 +1,7 @@
+package keeper_test
+
+import sdk "github.com/cosmos/cosmos-sdk/types"
+
+var (
+	InitTokens = sdk.TokensFromConsensusPower(200)
+)
